bug fix - program loading progress bar stuck
Fixed a bug where the program_loading progress meter get stuck in the
event of a failure during script initialization.  The was no way to
break out of the loading meter if an exception occurred in the wrong
place.  Fixed error handling so that it will kill the program loading
meter if the program crashes during initialization.