Increased efficiency
Fixed problem where customers with a large number of containers were
causing the script to hang if RANDOM set to True.  Created a
'RANDOM_CONTAINER_COUNT' that will be used to limit the number of cdn
enabled containers used for testing.  Creating the list is expensive
(time) and capping the list length increases performance of the script.